requirements.txt
# HullZero - Dependências Python

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Data Science
numpy==1.24.3
pandas==2.1.3
scipy==1.11.4

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.2
tensorflow==2.15.0
prophet==1.1.5

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
# timescaledb - Instalar separadamente se necessário (versões disponíveis: 0.0.1-0.0.4)
# timescaledb==0.0.4

# Caching
redis==5.0.1

# HTTP Clients
httpx==0.25.2
requests==2.31.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilities
python-dateutil==2.8.2
pytz==2023.3
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Code Quality
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8

